services/suspicious_action_detection/requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
pydantic
pydantic-settings